fix neoforge event constructor delegates
